# webui/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "My Proxy API",
    "description": "My Proxy provides an OpenAI-compatible API gateway that routes requests to configured providers. All admin endpoints require authentication via Bearer token or local/private IP access.\n\n## Authentication\n- **Admin API**: Bearer token (`Authorization: Bearer <api_key>`) or localhost access\n- **Public API** (`/v1/*`): Authenticated via the upstream provider's own API key (configured per-provider)",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/", "description": "Same-origin (proxy handles routing)" }
  ],
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "summary": "Chat Completions (OpenAI-compatible)",
        "description": "Send a chat completion request. Routes to the correct upstream provider based on the model name. Supports both streaming and non-streaming responses. Transparently translates to/from Gemini format when the target is a Gemini provider.",
        "tags": ["Public API"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": { "type": "string", "example": "gpt-4" },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string", "enum": ["system", "user", "assistant"] },
                        "content": { "type": "string" }
                      }
                    }
                  },
                  "stream": { "type": "boolean", "default": false },
                  "max_tokens": { "type": "integer" },
                  "temperature": { "type": "number" },
                  "top_p": { "type": "number" },
                  "top_k": { "type": "integer" }
                },
                "required": ["model", "messages"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful chat completion (non-streaming). Returns standard OpenAI chat completion response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "object": { "type": "string" },
                    "created": { "type": "integer" },
                    "model": { "type": "string" },
                    "choices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "index": { "type": "integer" },
                          "message": {
                            "type": "object",
                            "properties": {
                              "role": { "type": "string" },
                              "content": { "type": "string" }
                            }
                          },
                          "finish_reason": { "type": "string" }
                        }
                      }
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "prompt_tokens": { "type": "integer" },
                        "completion_tokens": { "type": "integer" },
                        "total_tokens": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/v1/embeddings": {
      "post": {
        "summary": "Create Embeddings (OpenAI-compatible)",
        "description": "Create embedding vectors for the given input text. Routes to the provider that owns the requested model.",
        "tags": ["Public API"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": { "type": "string" },
                  "input": { "type": "string" }
                },
                "required": ["model", "input"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful embedding response" },
          "400": { "description": "Bad request" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin API key obtained from the API Keys page. Not required when accessing from localhost."
      }
    },
...
  }
}